@@ -4,35 +4,43 @@
   "description": "",
   "main": "./lib/azusa.js",
   "scripts": {
-    "clean": "rimraf example",
-    "start": "webpack-dev-server --inline --hot --progress",
-    "dev": "npm run start",
-    "build": "webpack --config webpack.config.prod.js"
+    "clean": "rimraf lib && rimraf example",
+    "start": "webpack-dev-server --config webpack.config.js",
+    "build": "npm run build:m",
+    "build:m": "rimraf lib && webpack --config webpack.config.module.js",
+    "build:e": "rimraf example && webpack --config webpack.config.example.js && webpack-dev-server --config webpack.config.example.js"
   },
   "author": "EYHN",
   "license": "GPL-2.0",
   "devDependencies": {
-    "@types/node": "^7.0.22",
-    "@types/three": "^0.84.8",
-    "awesome-typescript-loader": "^3.1.3",
-    "babel-core": "^6.24.1",
-    "babel-loader": "^7.0.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-env": "^1.5.0",
-    "file-loader": "^0.11.1",
-    "html-webpack-plugin": "^2.28.0",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.0.0",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@types/node": "^10.9.4",
+    "@types/three": "^0.92.18",
+    "babel-loader": "^8.0.2",
+    "declaration-bundler-webpack-plugin": "^1.0.3",
+    "file-loader": "^2.0.0",
+    "fork-ts-checker-webpack-plugin": "^0.4.9",
+    "html-webpack-plugin": "^3.2.0",
     "rimraf": "^2.6.1",
-    "source-map-loader": "^0.2.1",
-    "typescript": "^2.3.3",
-    "uglify-js": "^3.0.10",
-    "url-loader": "^0.5.8",
-    "webpack": "^2.5.1",
-    "webpack-bundle-analyzer": "^2.8.2",
-    "webpack-dev-server": "^2.4.5"
+    "source-map-loader": "^0.2.4",
+    "ts-loader": "^5.1.0",
+    "typescript": "^3.0.3",
+    "uglify-js": "^3.4.9",
+    "url-loader": "^1.1.1",
+    "webpack": "^4.17.2",
+    "webpack-bundle-analyzer": "^2.13.1",
+    "webpack-cli": "^3.1.0",
+    "webpack-dev-server": "^3.1.8"
   },
   "dependencies": {
-    "three": "^0.85.2"
+    "three": "^0.96.0"
   },
-  "typings": "./lib/azusa.d.ts"
+  "types": "./lib/azusa.d.ts",
+  "typings": "./lib/azusa.d.ts",
+  "browserslist": [
+    "last 2 chrome version"
+  ]
 }
