@@ -1,8 +1,8 @@
 import Azusa from './azusa';
 
-const testSound = require("file-loader!./static/cha.mp3");
+import testSound from "./static/cha.mp3";
 
-const bgImg = require("file-loader!./static/9s.jpg");
+import bgImg from "./static/9s.jpg";
 
 const azusa = new Azusa({
   view: document.getElementById('app') as HTMLCanvasElement,
@@ -10,14 +10,18 @@ const azusa = new Azusa({
   cutEnd: 256
 });
 
-azusa.audio.load(testSound, undefined, (xhr: any) => {
-  console.log((xhr.loaded / xhr.total * 100) + '% loaded');
+azusa.audio.load({
+  src: testSound,
+  onPrgress: (xhr: ProgressEvent) => {
+    console.log((xhr.loaded / xhr.total * 100) + '% loaded');
+  }
 });
 
-azusa.audio.Volume = 0.5;
+azusa.audio.setVolume(0.5);
 
 window.addEventListener('resize', () => {
   azusa.resize(window.innerWidth, window.innerHeight);
 })
 
-document.getElementById('bg').style.backgroundImage = `url('${bgImg}')`
+const container = document.getElementById('bg')
+container && (container.style.backgroundImage = `url('${bgImg}')`)
