@@ -4,6 +4,12 @@ import * as THREE from 'three';
 export interface IAudioOption {
   fftsize?: number
 }
+export interface loadOption {
+  src: string,
+  onLoad?: (buffer: THREE.AudioBuffer) => void
+  onPrgress?: (xhr: ProgressEvent) => void
+  onError?: () => void
+}
 
 export class Audio extends EventEmitter {
   public listener: THREE.AudioListener;
@@ -17,21 +23,28 @@ export class Audio extends EventEmitter {
     this.sound = new THREE.Audio(this.listener);
     this.audioLoader = new THREE.AudioLoader();
     this.analyser = new THREE.AudioAnalyser(this.sound, option.fftsize || 256);
-    this.Volume = 0.5;
     this.frequencyBinCount = this.analyser.analyser.frequencyBinCount;
   }
-  load(src: string,onLoad?: Function, onPrgress?: Function, onError?: Function) {
-    this.audioLoader.load(src, (buffer: any) => {
+  public load (
+    option: loadOption
+  ) {
+    const {
+      src,
+      onLoad = (buffer: THREE.AudioBuffer) => void(0),
+      onPrgress = (xhr: ProgressEvent) => void(0),
+      onError = () => void(0)
+    } = option
+    this.audioLoader.load(src, (buffer: THREE.AudioBuffer) => {
       this.sound.setBuffer(buffer);
       this.sound.setLoop(true);
       this.sound.play();
       return onLoad(buffer);
     }, onPrgress, onError);
   }
-  set Volume(volume:number){
+  public setVolume (volume:number) {
     this.sound.setVolume(volume);
   }
-  getFrequencyData() {
+  public getFrequencyData () {
     return this.analyser.getFrequencyData()
   }
 }
